Flow.cpp, .h - Correct a sign bug in DistributeToReaches(). Fix a compiler warning. ReachRouting.cpp - In SolveReachKinematicWave(), remove the logic which adds the "externalFluxes" from GetReachFluxes() into the lateralInflow. My assumption is that we don't use the "external fluxes" part of the code at all now, and probably never did.  In GetLateralInflow(), flip the sign of the value returned by GetFluxValue().  In KinematicWave(), remove the assertion that the return value will be greater than zero. In ApplyReachOutflowWP(), when the kinematic wave solution is less than the minimum flow, use the minimum flow instead. Also, when there is a lateral withdrawal from the reach (as for irrigation or drinking water), set its temperature equal to the temperature in the reach after the runoff into the reach, the upstream inflow, and any magic water have been mixed in. WaterRights.cpp, .h - Add m_colReachXFLUX_D back to the AltWaterMaster class. Add XFLUX_D into the spring water accumulator in AltWaterMaster::EndStep().

+++ SourceCode/Flow/Flow.h
@@ -890,7 +890,7 @@ public:
       m_fluxWP.MixIn(H2OtoAddWP);
 
       float orig_stored_value = m_globalHandlerFluxValue;
-      m_globalHandlerFluxValue -= H2OtoAddWP.m_volume_m3; // Negative values of m_globalHandlerFluxValue represent water entering the reach.
+      m_globalHandlerFluxValue -= (float)H2OtoAddWP.m_volume_m3; // Negative values of m_globalHandlerFluxValue represent water entering the reach.
       if (isnan(m_globalHandlerFluxValue))
       {
          m_nanOccurred = true;
